Fix export
Export function before had the path of the old server within itself,
therefore it was not working in the container.

These changes still have hard coded paths, except this time:
- The paths are to do with the container, which should be more stable.
- The tabledata.zip file is stored in the /tmp/ dir instead of the
project folder.

The CSRF token is also sent to the server in the POST request, this must
have been due to a Django 2.0 change.